% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_plot.R
\name{coverage_plot}
\alias{coverage_plot}
\title{Plot a coverage simulation for a single estimator}
\usage{
coverage_plot(metasims_output)
}
\arguments{
\item{metasims_output}{A data.frame produced by \link{metasims}.}
}
\description{
Plot a coverage simulation for a single estimator
}
\seealso{
Other vis_tools: 
\code{\link{simpar_table}()},
\code{\link{zeta_plot}()}

Other reporting Functions and tools for reporting simulation results.: 
\code{\link{dist_name}()},
\code{\link{simpar_table}()},
\code{\link{zeta_plot}()}
}
\concept{reporting Functions and tools for reporting simulation results.}
\concept{vis_tools}
