% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_parameters.R
\docType{data}
\name{default_parameters}
\alias{default_parameters}
\title{Default Simulation Parameters}
\format{A dataframe with 6 rows and 2 variables.
\describe{
\item{dist}{R-friendly distribution name.}
\item{par}{Parameter set for r\code{dist} random sampling.}
}}
\usage{
default_parameters
}
\description{
These are the default distributional parameter sets used in \link{metasim}.
}
\seealso{
Other metapar Simulation metaparameter functions.: 
\code{\link{sim_df}()},
\code{\link{sim_dist}()}

Other neet_test_one One neet test has been written: 
\code{\link{beta_par}()},
\code{\link{coverage_plot}()},
\code{\link{dist_name}()},
\code{\link{intervention_proportion}()},
\code{\link{lr_se}()},
\code{\link{metamodel}()},
\code{\link{metasim}()},
\code{\link{metatrial}()},
\code{\link{plots}},
\code{\link{sim_df}()},
\code{\link{sim_dist}()},
\code{\link{sim_n}()}
}
\concept{metapar Simulation metaparameter functions.}
\concept{neet_test_one One neet test has been written}
\keyword{datasets}
