% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_label.R
\name{dist_label}
\alias{dist_label}
\title{Distribution label for plots}
\usage{
dist_label(dist, par)
}
\arguments{
\item{dist}{String specifying distribution, i.e., "norm".}

\item{par}{List specifying parameters for distribution.}
}
\description{
Convert a distribution and a list of parameters to a standard formatted
text distribution, with parameters rounded to two decimal points.
}
\details{
Works for distributions labelled by \link{dist_name}
}
