% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention_proportion.R
\name{intervention_proportion}
\alias{intervention_proportion}
\title{Calculate proportion of intervention group}
\usage{
intervention_proportion(n, proportion, error)
}
\arguments{
\item{n}{sample size}

\item{proportion}{Expected proportion for the intervention group}

\item{error}{Within what value will 90 per cent of the proportion
of intervention groups fall within?}
}
\description{
Calculate proportion of intervention group
}
\seealso{
Other simulation Functions that contribute to simulation pipeline.: 
\code{\link{beta_par}()},
\code{\link{default_parameters}},
\code{\link{sim_df}()},
\code{\link{sim_n}()},
\code{\link{sim_sample}()}

Other sample_size Generating meta-anlysis sample sizes.: 
\code{\link{beta_par}()},
\code{\link{sim_n}()}
}
\concept{sample_size Generating meta-anlysis sample sizes.}
\concept{simulation Functions that contribute to simulation pipeline.}
