% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotr_study.R
\docType{data}
\name{lotr_study}
\alias{lotr_study}
\title{Dummy set of studies}
\format{
An object of class \code{character} of length 383.
}
\usage{
lotr_study
}
\description{
These data were obtained from
\link{kaggle}(# https://www.kaggle.com/paultimothymooney/lord-of-the-rings-data/data).
}
\keyword{datasets}
