% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metasim.R
\name{metasim}
\alias{metasim}
\title{one row, one simulations}
\usage{
metasim(..., id = "simulation1", trial_fn = metatrial, trials = 4)
}
\arguments{
\item{...}{\code{trial_fn} arguments, i.e., simulation nparameters}

\item{trial_fn}{the function to repeat}

\item{trials}{the number of trials per simulation}
}
\description{
runs on one row, returns coverage probability
}
\seealso{
Other neet_test_one One neet test has been written: 
\code{\link{beta_par}()},
\code{\link{coverage_plot}()},
\code{\link{default_parameters}},
\code{\link{dist_name}()},
\code{\link{intervention_proportion}()},
\code{\link{lr_se}()},
\code{\link{metamodel}()},
\code{\link{metatrial}()},
\code{\link{plots}},
\code{\link{sim_df}()},
\code{\link{sim_dist}()},
\code{\link{sim_n}()}
}
\concept{neet_test_one One neet test has been written}
