% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_df.R
\name{sim_df}
\alias{sim_df}
\title{generate simulation parameter dataframe}
\usage{
sim_df(
  dist_df = default_parameters,
  k = c(3, 7, 20),
  tau_sq = seq(0, 0.4, by = 0.2),
  effect_ratio = c(1, 1.2, 1.5),
  min_n = 20,
  max_n = 200,
  prop = 0.5,
  prop_error = 0.1
)
}
\arguments{
\item{dist_df}{A dataframe.
with one column for distribution, and one column for the parameters.
Defaults to \link{default_parameters}. Note that the \code{par} arguments can be
changed, but only the distributions presented in \link{default_parameters} have
been implemented.}

\item{k}{Simulate for different numbers of studies.}

\item{tau_sq}{Variance \eqn{\gamma_k \sim N(0, \tau^2)} associated with the random effect}

\item{effect_ratio}{Ratio of population effects intervention / control}

\item{min_n}{lower bound for sample sizes, defaults to 20}

\item{max_n}{upper bound for sample sizes, defaults to 200}

\item{prop}{proportion of n we expect}

\item{prop_error}{what do we expect 90\% of proportions to fall within?}
}
\description{
generate simulation parameter dataframe
}
\seealso{
Other simulation Functions that contribute to simulation pipeline.: 
\code{\link{beta_par}()},
\code{\link{default_parameters}},
\code{\link{intervention_proportion}()},
\code{\link{sim_n}()},
\code{\link{sim_sample}()}
}
\concept{simulation Functions that contribute to simulation pipeline.}
