% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dist.R
\name{sim_dist}
\alias{sim_dist}
\title{Summarise sampling distributions}
\usage{
sim_dist(parameter_tibble, x_values = seq(0, 2, by = 0.001), output = "plot")
}
\arguments{
\item{parameter_tibble}{A tibble in the format of \link{default_parameters}, with
a column for distributions, R-friendly format ("norm", "lnorm"), and
parameters as a list-column.}

\item{output}{Specify "plot" or "table".}
}
\description{
Provides plot or table summary of distributions dampled from in \link{metasims}.
}
\examples{
sim_dist(default_parameters)
sim_dist(default_parameters, output = "table")

}
\seealso{
Other reporting Functions and tools for reporting simulation results.: 
\code{\link{coverage_plot}()},
\code{\link{dist_name}()},
\code{\link{metaplot}()},
\code{\link{simpar_table}()},
\code{\link{zeta_plot}()}

Other vis_tools: 
\code{\link{coverage_plot}()},
\code{\link{metaplot}()},
\code{\link{plots}},
\code{\link{simpar_table}()},
\code{\link{zeta_plot}()}

Other neet_test_one One neet test has been written: 
\code{\link{beta_par}()},
\code{\link{default_parameters}},
\code{\link{density_fn}()},
\code{\link{dist_name}()},
\code{\link{intervention_proportion}()},
\code{\link{lr_se}()},
\code{\link{metamodel}()},
\code{\link{metasims}()},
\code{\link{metasim}()},
\code{\link{metatrial}()},
\code{\link{plots}},
\code{\link{sim_df}()},
\code{\link{sim_n}()}
}
\concept{neet_test_one One neet test has been written}
\concept{reporting Functions and tools for reporting simulation results.}
\concept{vis_tools}
