% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_samples.R
\name{sim_samples}
\alias{sim_samples}
\title{Simulate samples}
\usage{
sim_samples(measure = "mean", measure_spread = "sd", sim_dat, trials = 3)
}
\arguments{
\item{measure}{Calculate sample median or mean. Defaults to mean.}

\item{measure_spread}{Defaults to standard deviation, \code{sd}. Specify "iqr", "range", "sd", "var". Will have to check how many of these are done right now.}

\item{sim_dat}{Dataframe created by \link{sim_df}.}

\item{trials}{Number of trials, that is repeated rows to return.
Defaults to 3 (small number of trials is good for testing).}
}
\description{
Function to repeat rows in sim_df and produces a sample for each
}
