% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_samples.R
\name{sim_trials}
\alias{sim_trials}
\title{Repeat rows}
\usage{
sim_trials(sim_dat, trials)
}
\arguments{
\item{sim_dat}{Dataframe created by \link{sim_df}.}

\item{trials}{Number of trials, that is repeated rows to return.
Defaults to 3 (small number of trials is good for testing).}
}
\description{
Repeat rows
}
