% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simeta_table_tex.R
\name{simeta_table_tex}
\alias{simeta_table_tex}
\title{Simeta table formating for Tex}
\usage{
simeta_table_tex(x, digits = 2, ...)
}
\arguments{
\item{x}{A dataframe to format for tex.}

\item{digits}{Defaults to 2.}

\item{...}{Arguments for \link{kable}.}
}
\description{
Simeta table formating for Tex
}
