% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpar_table.R
\name{simpar_table}
\alias{simpar_table}
\title{Table of simulation-level metaparameters}
\usage{
simpar_table(sim)
}
\arguments{
\item{sim}{Metasimulation object produced by \link{metasims}.}
}
\description{
Table of simulation-level metaparameters
}
\seealso{
Other vis_tools: 
\code{\link{coverage_plot}()},
\code{\link{zeta_plot}()}

Other reporting Functions and tools for reporting simulation results.: 
\code{\link{coverage_plot}()},
\code{\link{dist_name}()},
\code{\link{zeta_plot}()}
}
\concept{reporting Functions and tools for reporting simulation results.}
\concept{vis_tools}
