% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\docType{data}
\name{sims}
\alias{sims}
\title{Example simulation output}
\format{
An object of class \code{metasim} of length 4.
}
\usage{
sims
}
\description{
So I don't have to run sims every time.
}
\keyword{datasets}
