% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeta_plot.R
\name{zeta_plot}
\alias{zeta_plot}
\title{plot the distribution of the proportion allocated to the intervention group}
\usage{
zeta_plot(mu, epsilon)
}
\description{
plot the distribution of the proportion allocated to the intervention group
}
\seealso{
Other vis_tools: 
\code{\link{coverage_plot}()},
\code{\link{metaplot}()},
\code{\link{plots}},
\code{\link{sim_dist}()},
\code{\link{simpar_table}()}

Other reporting Functions and tools for reporting simulation results.: 
\code{\link{coverage_plot}()},
\code{\link{dist_name}()},
\code{\link{metaplot}()},
\code{\link{sim_dist}()},
\code{\link{simpar_table}()}
}
\concept{reporting Functions and tools for reporting simulation results.}
\concept{vis_tools}
