% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_par.R
\name{beta_par}
\alias{beta_par}
\title{Calculate distributional parametrs for proportion of sample size given
to intervention}
\usage{
beta_par(proportion, error)
}
\arguments{
\item{proportion}{Expected proportion for the intervention group}

\item{error}{Within what value will 90 per cent of the proportion
of intervention groups fall within?}
}
\description{
This function returns parameters for a beta distribution calculated from an
expected proportion assigned to intervention group and error.
}
\details{
Called by \link{intervention_proportion}, this function calculates
the parameters for the beta distribution to randomly generate
the proportion \eqn{p_k} of the total sample size
\eqn{N_k} for the
\eqn{k}th study.
}
\seealso{
Other sample_size Generating meta-anlysis sample sizes.: 
\code{\link{intervention_proportion}()},
\code{\link{sim_n}()}

Other neet_test_one One neet test has been written: 
\code{\link{default_parameters}},
\code{\link{density_fn}()},
\code{\link{dist_name}()},
\code{\link{intervention_proportion}()},
\code{\link{lr_se}()},
\code{\link{metamodel}()},
\code{\link{metasims}()},
\code{\link{metasim}()},
\code{\link{metatrial}()},
\code{\link{plots}},
\code{\link{sim_df}()},
\code{\link{sim_dist}()},
\code{\link{sim_n}()}

Other simulation Functions that contribute to simulation pipeline.: 
\code{\link{default_parameters}},
\code{\link{density_fn}()},
\code{\link{intervention_proportion}()},
\code{\link{lr_se}()},
\code{\link{metamodel}()},
\code{\link{metasims}()},
\code{\link{metasim}()},
\code{\link{metatrial}()},
\code{\link{sim_df}()},
\code{\link{sim_n}()},
\code{\link{sim_sample}()}
}
\concept{neet_test_one One neet test has been written}
\concept{sample_size Generating meta-anlysis sample sizes.}
\concept{simulation Functions that contribute to simulation pipeline.}
