% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage-plot.R
\name{coverage_plot}
\alias{coverage_plot}
\title{Plot a coverage simulation for a single estimator}
\usage{
coverage_plot(results_df)
}
\arguments{
\item{results_df}{A data.frame produced by \link{metasims}.}
}
\description{
Plot a coverage simulation for a single estimator
}
\concept{vis_tools}
