% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_plot.R
\name{coverage_plot}
\alias{coverage_plot}
\title{Plot a coverage simulation for a single estimator}
\usage{
coverage_plot(results_df)
}
\arguments{
\item{results_df}{A data.frame produced by \link{metasims}.}
}
\description{
Plot a coverage simulation for a single estimator
}
\seealso{
Other vis_tools: 
\code{\link{plots}},
\code{\link{sim_dist}()},
\code{\link{zeta_plot}()}

Other neet_test_one One neet test has been written: 
\code{\link{beta_par}()},
\code{\link{default_parameters}},
\code{\link{dist_name}()},
\code{\link{intervention_proportion}()},
\code{\link{lr_se}()},
\code{\link{metamodel}()},
\code{\link{metasim}()},
\code{\link{metatrial}()},
\code{\link{plots}},
\code{\link{sim_df}()},
\code{\link{sim_dist}()},
\code{\link{sim_n}()}
}
\concept{neet_test_one One neet test has been written}
\concept{vis_tools}
