% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_plot.R
\name{coverage_plot}
\alias{coverage_plot}
\title{Coverage Simulation Plot}
\usage{
coverage_plot(sims)
}
\arguments{
\item{sims}{A simulation object produced by \link{metasims}.}
}
\description{
This provides a coverage plot for a simulation of one estimator.
}
