% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_parameters.R
\docType{data}
\name{default_parameters}
\alias{default_parameters}
\title{Default Simulation Parameters}
\format{
A dataframe with 6 rows and 2 variables.
\describe{
\item{dist}{R-friendly distribution name.}
\item{par}{Parameter set for r\code{dist} random sampling.}
}
}
\usage{
default_parameters
}
\description{
These are the default distributional parameter sets used in \link{metasim}.
}
\seealso{
Other simulation Functions that contribute to simulation pipeline.: 
\code{\link{beta_par}()},
\code{\link{density_fn}()},
\code{\link{intervention_proportion}()},
\code{\link{lr_se}()},
\code{\link{metasims}()},
\code{\link{metasim}()},
\code{\link{metatrial}()},
\code{\link{sim_df}()},
\code{\link{sim_n}()},
\code{\link{sim_sample}()}

Other neet_test_one One neet test has been written: 
\code{\link{beta_par}()},
\code{\link{density_fn}()},
\code{\link{dist_name}()},
\code{\link{intervention_proportion}()},
\code{\link{lr_se}()},
\code{\link{metasims}()},
\code{\link{metasim}()},
\code{\link{metatrial}()},
\code{\link{sim_df}()},
\code{\link{sim_n}()}
}
\concept{neet_test_one One neet test has been written}
\concept{simulation Functions that contribute to simulation pipeline.}
\keyword{datasets}
