% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_fn.R
\name{density_fn}
\alias{density_fn}
\title{Get an arbitrary density function}
\usage{
density_fn(x, distribution, parameters, type = "q")
}
\arguments{
\item{x}{Input for density function.}

\item{distribution}{Currently programmed for "norm", "pareto", "exp", and
"lnorm" inputs.}

\item{parameters}{List of parameters for distribution; e.g. list(mean = 30,
sd = 0.2) for \eqn{normal(30, 0.2)}.}

\item{type}{Random "r" sample, density "d", quantile "q", or probability "p".}
}
\description{
Get an arbitrary density function
}
