% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotr_study.R
\docType{data}
\name{lotr_study}
\alias{lotr_study}
\title{Dummy set of studies}
\format{
An object of class \code{character} of length 383.
}
\usage{
lotr_study
}
\description{
Lord of the Rings characters combined with a random year.
}
\details{
The names data were obtained from
\href{https://www.kaggle.com/paultimothymooney/lord-of-the-rings-data/data}{kaggle}.

See data-raw/lotr_study.R to see how the dummy set of studies was created.
}
\keyword{datasets}
