% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamodel.R
\name{metamodel}
\alias{metamodel}
\title{Meta-analyse with error handling}
\usage{
metamodel(
  y = rnorm(10, mean = 20, sd = 0.2),
  se = runif(10, 0, 0.8),
  test = "knha",
  data = NULL
)
}
\arguments{
\item{test}{"knha" or "z"}
}
\description{
If random-effects model fails to converge, then replace with fixed-effects model,
else, produce error string.
}
\seealso{
Other neet_test_one One neet test has been written: 
\code{\link{beta_par}()},
\code{\link{default_parameters}},
\code{\link{density_fn}()},
\code{\link{dist_name}()},
\code{\link{intervention_proportion}()},
\code{\link{lr_se}()},
\code{\link{metasims}()},
\code{\link{metasim}()},
\code{\link{metatrial}()},
\code{\link{sim_df}()},
\code{\link{sim_n}()}

Other simulation Functions that contribute to simulation pipeline.: 
\code{\link{beta_par}()},
\code{\link{default_parameters}},
\code{\link{density_fn}()},
\code{\link{intervention_proportion}()},
\code{\link{lr_se}()},
\code{\link{metasims}()},
\code{\link{metasim}()},
\code{\link{metatrial}()},
\code{\link{sim_df}()},
\code{\link{sim_n}()},
\code{\link{sim_sample}()}
}
\concept{neet_test_one One neet test has been written}
\concept{simulation Functions that contribute to simulation pipeline.}
