% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{metaplot}
\alias{metaplot}
\title{Aggregate plots of simulation results}
\usage{
metaplot(sim, coverage = TRUE)
}
\arguments{
\item{sim}{Simulation results of the type produced by \link{metasims}.}
}
\description{
Aggregate plots of simulation results
}
\seealso{
Other vis_tools: 
\code{\link{coverage_plot}()},
\code{\link{plots}},
\code{\link{sim_dist}()},
\code{\link{simpar_table}()},
\code{\link{zeta_plot}()}

Other reporting Functions and tools for reporting simulation results.: 
\code{\link{coverage_plot}()},
\code{\link{dist_name}()},
\code{\link{sim_dist}()},
\code{\link{simpar_table}()},
\code{\link{zeta_plot}()}
}
\concept{reporting Functions and tools for reporting simulation results.}
\concept{vis_tools}
