% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-dist.R
\name{sim_dist}
\alias{sim_dist}
\title{Summarise sampling distributions}
\usage{
sim_dist(parameter_tibble, output = "plot")
}
\arguments{
\item{parameter_tibble}{A tibble in the format of \link{default_parameters}, with
a column for distributions, R-friendly format ("norm", "lnorm"), and
parameters as a list-column.}

\item{output}{Specify "plot" or "table".}
}
\description{
Provides plot or table summary of distributions dampled from in \link{metasims}.
}
