% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpar_table.R
\name{simpar_table}
\alias{simpar_table}
\title{Table of simulation-level metaparameters}
\usage{
simpar_table(sim)
}
\arguments{
\item{sim}{Metasimulation object produced by \link{metasims}.}
}
\description{
Table of simulation-level metaparameters
}
